+++ CHANGELOG.md
@@ -4,7 +4,13 @@
 
 ## [Unreleased]
 
+## [0.0.3]
+
+_2023-04-05_
+
 * Modify the `Image` fallback states background and icon colors #306
+* Integrate the new typography tokens #298
+* Add Legacy option to use previous DS style #310
 
 ## [0.0.2]
 
@@ -15,6 +21,11 @@ _2023-03-30_
 _2023-03-29_
 
 <!-- Links -->
-[Unreleased]: https://github.com/adevinta/spark-android/compare/0.0.2...HEAD
+
+[Unreleased]: https://github.com/adevinta/spark-android/compare/0.0.3...HEAD
+
+[0.0.3]: https://github.com/adevinta/spark-android/releases/tag/0.0.3
+
 [0.0.2]: https://github.com/adevinta/spark-android/releases/tag/0.0.2
+
 [0.0.1]: https://github.com/adevinta/spark-android/releases/tag/0.0.1